hcmut_lecturer_schedule-main/requirements.txt
Flask==3.0.0
Flask-CORS==3.0.10
google-auth>=2.0
google-auth-oauthlib>=0.4
google-api-python-client>=2.0
python-dotenv>=0.15
Flask-Session 
MarkupSafe>=2.1        
gunicorn==20.1.0
flasgger 